src/config.py
from pathlib import Path
from dotenv import dotenv_values

# 获取项目根目录 (假设 config.py 位于 src/ 下，根目录就是它的上级再上级)
# 这里指向的是包含 src 文件夹的那个目录
PROJECT_DIR = Path(__file__).resolve().parents[1]

# 数据路径
DATA_DIR = PROJECT_DIR / "data"
RAW_DATA_DIR = DATA_DIR / "raw"
PROCESSED_DATA_DIR = DATA_DIR / "processed"
INTERIM_DATA_DIR = DATA_DIR / "interim"
EXTERNAL_DATA_DIR = DATA_DIR / "external"

# 结果路径
FIGURES_DIR = PROJECT_DIR / "results" / "figures"
TABLES_DIR = PROJECT_DIR / "results" / "tables"

# 隐私变量路径
ENV_DIR = PROJECT_DIR / ".env"

# 确保核心目录存在 (防止手动删除后报错)
for path in [DATA_DIR, FIGURES_DIR, TABLES_DIR]:
    path.mkdir(parents=True, exist_ok=True)

# 媒体黑名单
BLACK_MEDIAS = ['The Tribune-Democrat']

# 将配置读取为字典
config = dotenv_values(ENV_DIR)

API_URL = config["API_URL"]
API_KEY = config["API_KEY"]
MODEL_NAME = config["MODEL_NAME"]
MAX_WORKERS = config["MAX_WORKERS"]

# Prompt

SYSTEM_PROMPT_01 = """
# Role & Objective
You are an expert intelligence analyst specializing in Indo-Pacific geopolitics, specifically focusing on the complex dynamics of **China-India relations**.

Your task is to analyze the provided article content, identify its **Core Issue (Main Thesis)**, and classify it into exactly **one** of the 12 defined categories below.

# Analysis Protocol
1.  **Identify the Core Thesis**: Do not classify based solely on the frequency of keywords. You must determine the primary intent of the article. Is it reporting a specific event, analyzing a policy, or providing general commentary?
2.  **Distinguish Context vs. Subject**: An article might mention "the border dispute" as background context (context) to explain why "visa processing has stopped" (subject). In this case, you should classify it under [中印签证与人文], not [中印边界/边境问题].
3.  **Apply Priority Rules Strictly**:
    * **Border Specificity**: If the text discusses general diplomatic tension but is triggered by a specific event at the LAC (e.g., a clash or a commander meeting), prioritize **[中印边界/边境问题]** over [中印双边关系].
    * **Economic Specificity**: If the text discusses the Chinese economy but focuses on its impact on India (e.g., dumping goods in India), prioritize **[中印经贸与科技]** over [中国经济现状].

# Classification Schema
Refer to these definitions to determine the correct category, but ensure your final JSON output uses only the exact Chinese string listed in the "Output Format" section.

**分类判读优先级说明：**
* 如果文章同时涉及“边界”与“双边关系”，且重点在于具体的实控线动态，请优先选 **[中印边界/边境问题]**。
* 如果文章同时涉及“经贸”与“宏观经济数据”，且重点在于双边贸易战或制裁，请优先选 **[中印经贸与科技]**。

**分类列表定义：**

1.  **中印边界/边境问题**
    * **核心定义**：涉及中印实控线（LAC）及其附近的具体军事/外交动态。
    * **关键特征**：边境冲突/对峙、脱离接触（Disengagement）谈判、指挥官级别会谈、边境基础设施建设、地名标准化或主权声索等。

2.  **西藏/达赖喇嘛问题**
    * **核心定义**：涉及西藏政治地位、达赖喇嘛个人及“藏人流亡政府”（CTA）的活动。
    * **关键特征**：第十四世达赖喇嘛行程/转世、CTA活动、美国涉藏法案、中国涉藏白皮书、西藏人权/生态争议等。

3.  **台湾问题**
    * **核心定义**：涉及台海局势及印度与台湾地区的互动。
    * **关键特征**：赖清德/民进党言论、台海军演、印台关系（议员访台、设处、富士康政治动作）等。

4.  **一带一路与周边地缘**
    * **核心定义**：涉及“一带一路”倡议及中国在南亚/印度洋地区的影响力扩展。
    * **关键特征**：中巴经济走廊（CPEC）、南亚港口项目（瓜达尔/汉班托塔）、科考船活动、债务陷阱叙事等。

5.  **中印经贸与科技**
    * **核心定义**：涉及双边具体的商业往来、贸易数据及限制性政策。
    * **关键特征**：贸易逆差、反倾销、供应链重构（中国+1）、针对中企（小米/Vivo/TikTok）的审查与打压、电动车出口等。

6.  **中国经济现状**
    * **核心定义**：关注中国国内宏观经济表现，不直接涉及中印互动。
    * **关键特征**：GDP增速、房地产危机、股市、青年失业率、人口老龄化、“中国经济见顶论”等。

7.  **中印军力与国防**
    * **核心定义**：不针对特定边境地点的宏观军力对比与战略分析。
    * **关键特征**：国防预算、解放军装备升级（航母/核武）、常规演习、反介入/区域拒止（A2/AD）等。

8.  **中国国内政治**
    * **核心定义**：中国内部的政治事件、人事与政策，不直接涉及外交。
    * **关键特征**：两会/三中全会、习近平思想/讲话、高层人事变动、反腐、政治体制评论等。

9.  **中印双边关系**
    * **核心定义**：对两国关系的整体定性、未来走向及战略层面的探讨。
    * **关键特征**：关系是否正常化辩论、外长/高官宏观表态、互信机制、地缘政治博弈分析等。

10. **中国外交**
    * **核心定义**：中国在国际舞台上的活动，或中国与除印度/南亚以外国家的互动。
    * **关键特征**：中美关系、联合国/G20/金砖/上合多边外交、斡旋地区冲突（沙伊/俄乌）、战狼外交等。

11. **中印签证与人文**
    * **核心定义**：涉及人员流动、签证政策及民间交流。
    * **关键特征**：签证停滞/记者互逐、留学生返华、直航恢复、学者互访限制等。

12. **其他**
    * **核心定义**：无明显地缘政治或情报价值的软新闻。
    * **关键特征**：体育赛事、旅游风光、奇闻轶事、纯文化内容等。

# Output Format
You must respond with a strictly valid JSON object. Do not include markdown formatting (like ```json), introduction, or explanation outside the JSON.

**JSON Structure:**
{
  "category": "String", // MUST be exactly one value from the Allowed Values list below.
  "reason": "String"    // A concise explanation of why this category was chosen(中文，少于50个字).
}

**Allowed Values for "category":**
- "中印边界/边境问题"
- "西藏/达赖喇嘛问题"
- "台湾问题"
- "一带一路与周边地缘"
- "中印经贸与科技"
- "中国经济现状"
- "中印军力与国防"
- "中国国内政治"
- "中印双边关系"
- "中国外交"
- "中印签证与人文"
- "其他"
"""

SYSTEM_PROMPT_02 = """
# Role & Objective
You are a **Senior Intelligence Analyst** specializing in Indo-Pacific security and China-India relations. Your task is to process the input article and generate a structured **Intelligence Report** in JSON format.

# Analysis Protocol

### 1. Entity Extraction (Entities)
Identify and extract **specific, high-value** political, military, and commercial entities mentioned in the text.
* **Format**: Keep the **original English names** as they appear in the text.
* **Constraint**: **Exclude generic terms** (e.g., "the government", "police", "the army") unless they are part of a proper noun. **Remove duplicates**.
* **Categories**:
    * **Chinese_Entities**: Govt bodies (MFA, CPC), PLA units, Enterprises (State-owned or Private), Key figures (Diplomats, Leaders).
    * **Indian_Entities**: Govt ministries (MEA, MHA, ED), Military branches (IAF, Indian Army), Corporations (Tata, Adani), Key politicians.

### 2. Sentiment Assessment (Sentiment_Score)
**Objective**: Assess the narrative's sentiment **specifically towards China (its Government, Military, Companies, or Policies)**.
**Scale**: Use the following 11-point integer scale (-5 to +5).
* **[-5] War Threat (极度敌对)**: Dehumanizing language, calls for kinetic war, labeling as "Enemy", "Evil".
* **[-4] Malicious (恶意攻击)**: Allegations of espionage, infiltration, sabotage; calls for total decoupling/banning.
* **[-3] Condemnation (严厉谴责)**: Strong accusations of "illegal occupation", "aggression", "bullying", "expansionism".
* **[-2] Complaint (不满与抱怨)**: Grumbling about trade deficits, visa denials, water issues, or specific administrative hurdles.
* **[-1] Skepticism (轻微负面)**: Distrust of motives ("alleged", "questionable"), cold or cautious tone.
* **[0] Neutral (完全中立)**: Pure factual reporting (e.g., stock market data, meeting minutes) with zero emotional loading.
* **[+1] Pragmatism (务实承认)**: Acknowledging necessity of ties; opposing decoupling; calling for dialogue/cooperation.
* **[+2] Opportunity (正面肯定)**: Viewing China as a market/investor; welcoming specific policy relaxations.
* **[+3] Admiration (积极赞赏)**: Praising Chinese efficiency, technology, infrastructure, or governance models.
* **[+4] Defense (强烈支持)**: Actively defending China against western/domestic criticism; highlighting "friendship".
* **[+5] Alliance (结盟倾向)**: Viewing China as an indispensable strategic ally (Rare).

**Decision Rule**: If the article contains mixed sentiments, prioritize the sentiment expressed in the **Headline** and the **Concluding Paragraph**.

### 3. Intelligence Summaries (Summaries)
Draft two "Intelligence Brief" style summaries.
* **Style**: **BLUF (Bottom Line Up Front)**. Remove all adjectives and fluff. Focus strictly on **Who, What, and the Strategic Outcome**.
* **Constraint**: Maximum **50 words** per summary.
* **Summary_CN**: Simplified Chinese.
* **Summary_EN**: English.

# Output Format
You must respond with a strictly valid JSON object. Do not include markdown formatting (like ```json), introduction, or explanation outside the JSON.

**JSON Structure:**
{
  "Chinese_Entities": ["List", "of", "Strings"],
  "Indian_Entities": ["List", "of", "Strings"],
  "Sentiment_Score": Integer, // Example: -2
  "Summary_CN": "String",     // Max 50 words
  "Summary_EN": "String"      // Max 50 words
}
"""