pyproject.toml
[project]
name = "public-opinion-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyterlab>=4.5.0",
    "matplotlib>=3.10.7",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
